moved error_handlers to their own test file
